refactor: split packing materials page into settings tab + tab shell
