+++ frontend/src/components/packing-materials/ConsumptionHistoryTab.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const ConsumptionHistoryTab: React.FC = () => <div />;
+
+export default ConsumptionHistoryTab;
